refactor hightlighting namespace
