+++ Common/UserInterface/HighlightingRules/HighlightRules.xsd
