+++ Common/UserInterface/HighlightingRules/HighlightRules.xshd
